Fix registries that rely on order
Vanilla no longer allows to register items to arbitrary
ids but some of our registries rely on this due to the
way vanilla stores this data and has no natural registries
for them.

Fixing this by registering the items in the order we want
them to be present in and then verying their ids match the
ones we expected afterwards.

While this is very brittle, it is the most straight forward
way to ensure we stay in sync with vanilla elements
without more complex changes.